# swagger/openapi.yaml
openapi: "3.1.0"
info:
  title: Registration API
  version: 1.0.0
servers:
  - url: http://backend.vm1.test
    description: VM1 backend
  - url: http://backend.vm2.test
    description: VM2 backend
  - url: http://backend.vm3.test
    description: VM3 backend
  - url: http://backend.vm4.test
    description: VM4 backend
  - url: http://backend.vm5.test
    description: VM5 backend
paths:
...
  /api/registrations/{registration}:
    get:
      summary: Megjelenít egyetlen regisztrációt
      description: |
         Keresse ki a megfelelő rekordot az adatbázisból az `id` alapján.
      operationId: registrations.show
      tags:
        - Registration
      parameters:
        - name: registration
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Egy regisztráció megjelenítése
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Registration'
        '404':
          description: Az elem nem található
    delete:
      summary: Regisztráció törlése
      description: |
         Törölje ki a megfelelő rekordot az adatbázisból az `id` alapján.
      operationId: registrations.destroy
      tags:
        - Registration
      parameters:
        - name: registration
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Sikeres törlés
        '404':
          description: Az elem nem található
        '500':
          description: Szerver hiba
...
components:
  schemas:
    Registration:
      type: object
      properties:
        id:
          type: integer
          description: A regisztráció azonosítója
        name:
          type: string
          maxLength: 40
          description: A regisztráló neve
        vegetarian:
          type: boolean
          description: Vegetáriánus étrendet kér -e, vagy sem
        date:
          type: string
          format: date
          description: Erre a napra szól a regisztráció
        arrived:
          type: string
          format: time
          nullable: true
          description: Az érkezés ideje, ha nem érkezett meg, akkor null